+++ @planetarium/account-web3-secret-storage/src/index.ts
@@ -2,3 +2,4 @@ export { KeyId } from "./KeyId.js";
 export { PassphraseEntry } from "./PassphraseEntry.js";
 export { TtyPassphraseEntry } from "./TtyPassphraseEntry.js";
 export { getDefaultWeb3KeyStorePath, Web3KeyStore } from "./Web3KeyStore.js";
+export { Web3Account, type Web3KeyObject } from "./Web3Account.js";