+++ @planetarium/account/src/RawPrivateKey.ts
@@ -16,7 +16,7 @@ export class RawPrivateKey {
       throw new Error(`Expected Uint8Array, but got ${typeof bytes}`);
     } else if (bytes.length !== 32) {
       throw new Error(
-        `Incorrect private key length; expected 32 bytes, but got ${bytes.length} bytes`
+        `Incorrect private key length; expected 32 bytes, but got ${bytes.length} bytes`,
       );
     } else if (!secp256k1.utils.isValidPrivateKey(bytes)) {
       throw new Error("Invalid private key");
@@ -32,7 +32,7 @@ export class RawPrivateKey {
       throw new Error(`Expected string, but got ${typeof hex}`);
     } else if (hex.length !== 64) {
       throw new Error(
-        `Incorrect private key length; expected 64 hexadigits, but got ${hex.length} hexadigits`
+        `Incorrect private key length; expected 64 hexadigits, but got ${hex.length} hexadigits`,
       );
     }
     const bytes = new Uint8Array(Buffer.from(hex, "hex"));
@@ -47,22 +47,31 @@ export class RawPrivateKey {
     return this.fromBytes(secp256k1.utils.randomPrivateKey());
   }
 
+  /**
+   * @deprecated Use {@link getPublicKey()} instead.
+   */
   get publicKey(): PublicKey {
+    // TODO: This attribute is deprecated.  We should remove it and make
+    // getPublicKey() method the only choice in the future.
     if (typeof this.#publicPart === "undefined") {
       this.#publicPart = PublicKey.fromBytes(
         secp256k1.getPublicKey(this.#privatePart),
-        "uncompressed"
+        "uncompressed",
       );
     }
 
     return this.#publicPart;
   }
 
+  getPublicKey(): Promise<PublicKey> {
+    return Promise.resolve(this.publicKey);
+  }
+
   async sign(message: Message): Promise<Signature> {
     const sig = await secp256k1.sign(
       await hashMessage(message),
       this.#privatePart,
-      { der: true }
+      { der: true },
     );
     return Signature.fromBytes(sig);
   }